requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-wtf==1.2.1
flask-mail==0.10.0
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
werkzeug==3.0.3
SQLAlchemy==2.0.31
ta==0.11.0
feedparser==6.0.11
anthropic==0.28.0
gunicorn==22.0.0
wtforms==3.1.2
email-validator==2.2.0
